fix: refactor values
