+++ src/types/formValues.ts
@@ -1,0 +1,12 @@
+export type FilterValues = {
+  search: string;
+  options: string[];
+  form: string;
+};
+
+export type BookFormValues = {
+  name: string;
+  email: string;
+  dates: [Date | null, Date | null];
+  comment: string;
+};
